[fix]: 대량 객체 생성 기능 버그 수정 및 파일 분리화 - 렌더링 부분 Object, Grid, Pattern 별 컴포넌트 파일 분리 - 생성된 객체 안에 텍스트 수정 시도 시 객체가 날아가버리는 버그 수정
